requirements.txt
pandas
pyarrow
requests
google-cloud-bigquery
pandas-gbq
dbt-bigquery